refactor: Improve Alert component and add redirect functionality
