+++ frontend/src/pages/budgetLines/CreateBudgetLine.jsx
@@ -1,7 +1,7 @@
 import CreateBudgetLineFlow from "./CreateBudgetLineFlow";
 import StepSelectProjectAndAgreement from "./StepSelectProjectAndAgreement";
 import StepCreateBudgetLines from "../../components/BudgetLineItems/CreateBLIsAndSCs";
-import StepSuccess from "./StepSuccess";
+import StepSuccessAlert from "../../components/UI/Alert/StepSuccessAlert";
 import { useBudgetLines } from "./budgetLineContext";
 
 export const CreateBudgetLine = () => {
@@ -22,10 +22,14 @@ export const CreateBudgetLine = () => {
                 selectedResearchProject={selectedProject}
                 selectedAgreement={selectedAgreement}
                 selectedProcurementShop={selectedProcurementShop}
-                existingBudgetLines={existingBudgetLines}
+                budgetLines={existingBudgetLines}
                 continueBtnText="Create Budget Lines"
             />
-            <StepSuccess />
+            <StepSuccessAlert
+                heading="Budget Lines Created"
+                message="The budget lines have been successfully created. You will be redirected to the Agreements list page."
+                link="/agreements"
+            />
         </CreateBudgetLineFlow>
     );
 };
